Update several more uses of internal
