+++ core/src/test/kotlin/net/corda/core/transactions/LedgerTransactionQueryTests.kt
@@ -10,6 +10,7 @@ import net.corda.core.identity.Party
 import net.corda.node.services.api.IdentityServiceInternal
 import net.corda.testing.contracts.DummyContract
 import net.corda.testing.core.*
+import net.corda.testing.internal.MockCordappProvider
 import net.corda.testing.internal.rigorousMock
 import net.corda.testing.node.MockServices
 import org.junit.Before
@@ -36,7 +37,7 @@ class LedgerTransactionQueryTests {
 
     @Before
     fun setup() {
-        services.mockCordappProvider.addMockCordapp(DummyContract.PROGRAM_ID, services.attachments)
+        services.addMockCordapp(DummyContract.PROGRAM_ID, services.attachments)
     }
 
     interface Commands {
