+++ testing/test-utils/src/main/kotlin/net/corda/testing/core/SerializationTestHelpers.kt
@@ -16,6 +16,8 @@ import net.corda.nodeapi.internal.serialization.*
 import net.corda.nodeapi.internal.serialization.amqp.AMQPClientSerializationScheme
 import net.corda.nodeapi.internal.serialization.amqp.AMQPServerSerializationScheme
 import net.corda.testing.common.internal.asContextEnv
+import net.corda.testing.internal.createTestSerializationEnv
+import net.corda.testing.internal.inVMExecutors
 import net.corda.testing.internal.rigorousMock
 import net.corda.testing.internal.testThreadFactory
 import org.apache.activemq.artemis.core.remoting.impl.invm.InVMConnector
@@ -25,8 +27,6 @@ import org.junit.runners.model.Statement
 import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.ExecutorService
 import java.util.concurrent.Executors
-
-private val inVMExecutors = ConcurrentHashMap<SerializationEnvironment, ExecutorService>()
 
 /** @param inheritable whether new threads inherit the environment, use sparingly. */
 class SerializationEnvironmentRule(private val inheritable: Boolean = false) : TestRule {
@@ -78,42 +78,3 @@ interface GlobalSerializationEnvironment : SerializationEnvironment {
     fun unset()
 }
 
-/**
- * Should only be used by Driver and MockNode.
- * @param armed true to install, false to do nothing and return a dummy env.
- */
-fun setGlobalSerialization(armed: Boolean): GlobalSerializationEnvironment {
-    return if (armed) {
-        object : GlobalSerializationEnvironment, SerializationEnvironment by createTestSerializationEnv("<global>") {
-            override fun unset() {
-                _globalSerializationEnv.set(null)
-                inVMExecutors.remove(this)
-            }
-        }.also {
-            _globalSerializationEnv.set(it)
-        }
-    } else {
-        rigorousMock<GlobalSerializationEnvironment>().also {
-            doNothing().whenever(it).unset()
-        }
-    }
-}
-
-private fun createTestSerializationEnv(label: String): SerializationEnvironmentImpl {
-    val factory = SerializationFactoryImpl().apply {
-        registerScheme(KryoClientSerializationScheme())
-        registerScheme(KryoServerSerializationScheme())
-        registerScheme(AMQPClientSerializationScheme(emptyList()))
-        registerScheme(AMQPServerSerializationScheme(emptyList()))
-    }
-    return object : SerializationEnvironmentImpl(
-            factory,
-            AMQP_P2P_CONTEXT,
-            KRYO_RPC_SERVER_CONTEXT,
-            KRYO_RPC_CLIENT_CONTEXT,
-            AMQP_STORAGE_CONTEXT,
-            KRYO_CHECKPOINT_CONTEXT
-    ) {
-        override fun toString() = "testSerializationEnv($label)"
-    }
-}
