+++ testing/test-utils/src/main/kotlin/net/corda/testing/dsl/TestDSL.kt
@@ -14,7 +14,7 @@ import net.corda.core.transactions.SignedTransaction
 import net.corda.core.transactions.TransactionBuilder
 import net.corda.core.transactions.WireTransaction
 import net.corda.testing.services.MockAttachmentStorage
-import net.corda.testing.services.MockCordappProvider
+import net.corda.testing.internal.MockCordappProvider
 import net.corda.testing.core.dummyCommand
 import java.io.InputStream
 import java.security.PublicKey
